+++ index.js
@@ -194,16 +194,12 @@ module.exports = {
 
         // Adding event listeners to take automatic screenshot
         if (autoCaptureOptions.indexOf('click') !== -1) {
-            flow.on(scheduleTask, function (task) {
-                driver.getSession() && driver.getSession().then(function (session) {
-                    if (session && task !== undefined && task.indexOf('WebElement.click') !== -1) {
-                        var filename = 'ScreenShot_onClick-' + process.pid + '-' + new Date().getTime();
-                        flow.wait(function () {
-                            return nemo.screenshot.snap(filename);
-                        }, 10000);
-                    }
-                });
-            });
+            var oclick = nemo.wd.WebElement.prototype.click;
+            nemo.wd.WebElement.prototype.click = function () {
+                var filename = 'ScreenShot_onClick-' + process.pid + '-' + new Date().getTime();
+                return nemo.screenshot.snap(filename)
+                    .then(oclick.bind(this));
+            };
         }
 
         if (autoCaptureOptions.indexOf('exception') !== -1) {